support automatically installing an existing flake config from bootstrap
previously, setting up nix-on-droid to use an existing flake config
meant waiting for the bootstrap to switch to the flake template you're
not going to use, deleting the flake it created, and then switching to
your own flake.

this allows building a bootstrap zip that automatically switches to a
flake specified in the CONFIG_FLAKE_URL environment variable, skipping
the flake template.